requirements.txt
tensorflow>=2.8.0
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.21.0
pillow>=8.0.0
scikit-learn>=0.24.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0 